+++ bflibcpp/makefile
@@ -24,7 +24,7 @@ CPPFLAGS += -std=c++20
 #
 # All objects defined in headers should be copied to 
 # bin
-CPP_OBJECTS = file string time path directory bfobject
+CPP_OBJECTS = file string time path directory object
 
 ### RELEASE ###
 R_CPPFLAGS = $(CPPFLAGS) -Ibin/release/ -I../bflibc/bin/release $(BF_LIB_C_UUID_FLAGS)
